Move the dev dynamic route builder into its own function
In dev the router matches a request against a list of dynamic routes once an
exact path lookup misses. That list is built inline in the watchpack handler in
setup-dev-bundler.ts, in the middle of a long function that does a lot of other
things.

This moves it out into buildDevDynamicRoutes. No behavior change: the data
routes still come first, so a data request doesn't match the page route it
shadows, and the caller still passes in the sorted routes it already computed.

A PR stacked on top of this one builds the same list from the routes Turbopack
reports, so it needs a second caller.
